% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_df.R
\name{labels_df}
\alias{labels_df}
\title{Create a data frame of the variable labels}
\usage{
labels_df(df)
}
\arguments{
\item{df}{A data frame imported from SPSS using `haven::read_sav()`}
}
\value{
A data frame with 2 columns, listing the names of the variables and the corresponding labels from the metadata.
}
\description{
Create a data frame of the variable labels
}
\examples{
gss_data <- haven::read_sav(system.file("extdata", "GSS2016.sav", package = "havenmisc"))
labels_df(gss_data)
}
